+++ src/components/pages/apps/matcalc/history/history-provider/history-provider.tsx
@@ -104,7 +104,7 @@ export function HistoryProvider({ children }: IChildrenProps) {
         plots,
         groupRows,
         genesets,
-        //groupsName,
+
         mode,
       })
     },