+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Gsea bubble now uses portal for tooltips.
 - IAxisConfig now exists for configuring axes through props.
 - History dev panel now updates when plots are added.
 - Refactored history to split actions into separate files for managability.